+++ exporters/elasticsearch/include/opentelemetry/exporters/elasticsearch/es_log_recordable.h
@@ -97,14 +97,6 @@ public:
   void SetTraceFlags(const opentelemetry::trace::TraceFlags &trace_flags) noexcept override;
 
   /**
-   * Set an attribute of a log.
-   * @param key the name of the attribute
-   * @param value the attribute value
-   */
-  void SetAttribute(nostd::string_view key,
-                    const opentelemetry::common::AttributeValue &value) noexcept override;
-
-  /**
    * Set Resource of this log
    * @param Resource the resource to set
    */
@@ -116,6 +108,15 @@ public:
    */
   void SetInstrumentationScope(const opentelemetry::sdk::instrumentationscope::InstrumentationScope
                                    &instrumentation_scope) noexcept override;
+
+protected:
+  /**
+   * Set an attribute of a log.
+   * @param key the name of the attribute
+   * @param value the attribute value
+   */
+  void SetAttributeImpl(nostd::string_view key,
+                        const opentelemetry::common::AttributeValue &value) noexcept override;
 
 private:
   // Define a JSON object that will be populated with the log data