+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_log_recordable.h
@@ -89,14 +89,6 @@ public:
   void SetTraceFlags(const opentelemetry::trace::TraceFlags &trace_flags) noexcept override;
 
   /**
-   * Set an attribute of a log.
-   * @param key the name of the attribute
-   * @param value the attribute value
-   */
-  void SetAttribute(nostd::string_view key,
-                    const opentelemetry::common::AttributeValue &value) noexcept override;
-
-  /**
    * Set Resource of this log
    * @param Resource the resource to set
    */
@@ -108,6 +100,15 @@ public:
    */
   void SetInstrumentationScope(const opentelemetry::sdk::instrumentationscope::InstrumentationScope
                                    &instrumentation_scope) noexcept override;
+
+protected:
+  /**
+   * Set an attribute of a log.
+   * @param key the name of the attribute
+   * @param value the attribute value
+   */
+  void SetAttributeImpl(nostd::string_view key,
+                        const opentelemetry::common::AttributeValue &value) noexcept override;
 
 private:
   proto::logs::v1::LogRecord proto_record_;